# apps/service-gateway/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.SysProductionOrder:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      itemCode:
        type: string
      orderNo:
        type: string
      plannedQty:
        type: number
      status:
        type: string
      updatedAt:
        type: string
    type: object
  models.SysSalesOrder:
    properties:
      bpID:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      docNo:
        type: string
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      lines:
        items:
          $ref: '#/definitions/models.SysSalesOrderLine'
        type: array
      totalAmount:
        type: number
      updatedAt:
        type: string
    type: object
  models.SysSalesOrderLine:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      docID:
        type: integer
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      itemCode:
        type: string
      qty:
        type: number
      unitPrice:
        type: number
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 这是一个高度可扩展的开源 ERP 底座，支持动态字段(JSONB)、复杂审批流以及物料需求计划(MRP)。
  title: OpenERP 企业级信息化中枢 API
  version: "1.0"
paths:
...
  /openerp/v1/production-orders:
    get:
      consumes:
      - application/json
      description: 获取生产订单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取生产订单列表
      tags:
      - 生产管理
    post:
      consumes:
      - application/json
      description: 创建一个新的生产订单
      parameters:
      - description: 生产订单数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.SysProductionOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建生产订单
      tags:
      - 生产管理
  /openerp/v1/production-orders/{id}:
    delete:
      consumes:
      - application/json
      description: 删除生产订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除生产订单
      tags:
      - 生产管理
    put:
      consumes:
      - application/json
      description: 更新生产订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 生产订单数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.SysProductionOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新生产订单
      tags:
      - 生产管理
  /openerp/v1/sales-orders:
    get:
      consumes:
      - application/json
      description: 获取包含行项目的销售订单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取销售订单列表
      tags:
      - 销售管理
    post:
      consumes:
      - application/json
      description: 创建一个新的销售订单及其行项目
      parameters:
      - description: 销售订单数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.SysSalesOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建销售订单
      tags:
      - 销售管理
  /openerp/v1/sales-orders/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除销售订单（及其行项目）
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除销售订单
      tags:
      - 销售管理
    put:
      consumes:
      - application/json
      description: 更新销售订单及其关联项目
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 销售订单数据
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.SysSalesOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新销售订单
      tags:
      - 销售管理
swagger: "2.0"